[FIX]scrollAnimation수정
